+++ src/rules/utils/accessors.ts
@@ -58,8 +58,7 @@ const isTemplateLiteral = <V extends string>(
   (value === undefined || node.quasis[0].value.raw === value);
 
 export type StringNode<S extends string = string> =
-  | StringLiteral<S>
-  | TemplateLiteral<S>;
+  StringLiteral<S> | TemplateLiteral<S>;
 
 /**
  * Checks if the given `node` is a {@link StringNode}.
@@ -165,5 +164,4 @@ export const getAccessorValue = <S extends string = string>(
     : getStringValue(accessor);
 
 export type AccessorNode<Specifics extends string = string> =
-  | StringNode<Specifics>
-  | KnownIdentifier<Specifics>;
+  StringNode<Specifics> | KnownIdentifier<Specifics>;